fix: address ask projection review findings
